loader obj v1
attention erreur de linkage pour tiny_obj_loader

+++ Engine/Engine/GUI.h
@@ -15,8 +15,8 @@ public:
 private:
 	//function
 	void TwMouseButtonBinder(GLFWwindow* window, int button, int action, int mods) { TwEventMouseButtonGLFW(button, action); }
-	void TwMouseMotionBinder(GLFWwindow* window, double x, double y) { TwEventMousePosGLFW(x, y); }
-	void TwMouseWheelBinder(GLFWwindow* window, double xoffset, double yoffset) { TwEventMouseWheelGLFW(xoffset); }
+	void TwMouseMotionBinder(GLFWwindow* window, double x, double y) { TwEventMousePosGLFW((int)x, (int)y); }
+	void TwMouseWheelBinder(GLFWwindow* window, double xoffset, double yoffset) { TwEventMouseWheelGLFW((int)xoffset); }
 	void TwKeyboardBinder(GLFWwindow* window, int key, int scancode, int action, int mods) { TwEventKeyGLFW(key,action); }
 
 	TwBar*	objTweakBar;
